+++ react-nd/src/App.jsx
@@ -1,11 +1,15 @@
+import { StrictMode } from "react";
+import { RouterProvider } from "react-router-dom";
 import { UserProvider } from "@/context/UserContext";
-import Routes from "@/Routes"
+import { router } from "@/Router";
 
 const App = () => {
   return (
-    <UserProvider>
-      <Routes />
-    </UserProvider>
+    <StrictMode>
+      <UserProvider>
+        <RouterProvider router={router} />
+      </UserProvider>
+    </StrictMode>
   );
 };
