+++ react-nd/src/views/category/index.js
@@ -1,0 +1,4 @@
+export * from "@/views/category/Category";
+export * from "@/views/category/CategorySidebar";
+export * from "@/views/category/CategoryMain";
+export * from "@/views/category/BusinessCard";
