+++ react-nd/src/views/error/index.js
@@ -1,0 +1,1 @@
+export * from "@/views/error/ErrorPage";
