+++ react-nd/src/views/home/index.js
@@ -1,0 +1,5 @@
+export * from "@/views/home/Home";
+export * from "@/views/home/HomeHero";
+export * from "@/views/home/CategoryList";
+export * from "@/views/home/ServiceCard";
+export * from "@/views/home/SearchBar";
